.mdai/requirements.txt
numpy==1.18.1
opencv-python==4.2.0.32
pillow==7.0.0
pydicom==1.4.1
scipy==1.4.1
torch==1.4.0
matplotlib==3.2.1
